wifi: Check YOYO_MAGIC fw dump validity
The structure of the Intel YOYO_MAGIC fw dump (All WiFi chips but StP2)
is different from Memory API MAGIC. The expected files are not always the
same because it depends on a parameter called Time point which is a
condition which the driver can identify and trigger a dump upon.

To know the expected files in the YOYO_MAGIC fw dump, the
ExpectedRegionID in the generated file dump_info.lst must be read. The
ExpectedRegionID has the Region IDs for all the expected files in the fw
dump. For more information refer to http://b/169152720#comment78.

Changed the script to read the expected region IDs in dump_info.lst.
Then check that files with these region IDs exist in the YOYO_MAGIC fw
dump.

BUG=b:169152720
TEST=Ran the test in two DUTs:
With YOYO MAGIC (Koaku):
     tast run $DUT wifi.CheckIntelFWDump [[ PASS ]]
With non Memory API MAGIC (Eve-StP2):
     tast run $DUT wifi.CheckIntelFWDump [[ PASS ]]

Change-Id: I4404a87d44e7dda1c063a514bb052d84a6c5b143
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3307206
Tested-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Joshua Emele <jemele@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>