camera: Use chrome-unstrusted://media-app for CCAUIGalleryButton
Sometimes CCAUIGalleryButton will hit a race condition when we try to
close the media app through CDP and the media app is still alive.

It is because there is an underlying issue when a navigation to media
app happens between closing it. And using chrome-untrusted://media-app
can avoid such race.

Once the underlying issue is solved, we can revert this workaround.

BUG=b:204528998
TEST=tast run [DUT] camera.CCAUIGalleryButton

Change-Id: I4bcf7772cae6a6e6d86c90274efa66332539c5ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3308574
Tested-by: Wei Lee <wtlee@chromium.org>
Auto-Submit: Wei Lee <wtlee@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>