camera: Use chrome-unstrusted://media-app for CCAUIGalleryButton
Sometimes CCAUIGalleryButton will hit a race condition that when we
try to close the media app through CDP, the media app is still alive.

It is because there is an underlying issue that when a navigation
happens between closing the media app. And using
chrome-untrusted://media-app can avoid such race.

Once the underlying issue is solved, we can revert this workaround.

BUG=b:204528998
TEST=tast run [DUT] camera.CCAUIGalleryButton

Change-Id: I4bcf7772cae6a6e6d86c90274efa66332539c5ef