Always place logfile at the same location for lacros.
Currently, lacros launched via shelf and via command line had different
locations for the lacros logfile. This makes it difficult to copy the
logfile when taking a faillog. It also means that the shelf lacros
logfile isn't put into the test out directory unless it is explicitly
copied.

This change puts the lacros log file always in the same location, and
copies it at the end of tests using the lacros fixture.

BUG=chromium:1263337
TEST=tast run <dut> lacros.Basic lacros.ShelfLaunch

Change-Id: Ib5b5dd0543d94ebf084ed7803e1abcd0c178e5df