video: Verify decoded frames by md5 checksums in video.WebCodecsDecode
This CL add frame content validation by comparing md5 checksums in
video.WebCodecsDecode. However, if the format of decoded frames are
RGB format, they are likely converted to the RGB format using GPU
after decoding. So their contents are dependent on GPU. We skip the
md5sum validation in the case.

BUG=b:173164490
TEST=video.WebCodecsDecode.*

Change-Id: I693de5fb1758c5101f7b0a7e53c116844dfa8e9c