cdm-oemcrypto: Filter out Huge OEMCrypto tests
For v16 of OEMCrypto there were a lot of unit tests added that validate
huge buffers or number of samples that can take a long time to execute.
Filter these tests out, which is the same thing Widevine does in their
automated testing environment. We will still run them manually when we
do updates though.

BUG=b:197688306
TEST=tast run DUT video.CDMOEMCrypto.oemcrypto (ps shows correct filter
being applied on DUT)

Change-Id: Ia3b69faded1c73ce33926aac5a27e2ed67c5d858