tast-tests: Update Crostini setup to match 'Turn on' button in settings.
In preparation for crrev.com/c/3312662, which updates the accessibility
label for Crostini's 'Turn on' button, this CL updates the Tast code to
match the new label. Unfortunately, this label is not Crostini-specific
and there are no Crostini labels we can match against, but at least for
now the label 'Turn on' is unique.

BUG=chromium:1227818
TEST=tast run $IP crostini.Launcher.system_wayland_buster_stable, both
with and without crrev.com/c/3312662

Change-Id: I1cf2be54a56d12620e8fc560aea8f17cc7c2de9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3312664
Tested-by: Timothy Loh <timloh@chromium.org>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>