faft: Convert firmware_SoftwareSync to tast
Add new function to BiosService to corrupt an EC section.  It adds one
to every byte of the section.

Add new test firmware.SoftwareSync. It has mostly the same functionality
as the autotest, but doesn't handle the EFS2 cases yet, and some steps
were omitted until they can be checked to make sure they are really
needed.

This test also clears DISABLE_EC_SOFTWARE_SYNC instead of failing like
the old test did. Since it restores the firmware backup, if a custom EC
was being used with DISABLE_EC_SOFTWARE_SYNC, it should back in that
state after the test.

BUG=b:194910957
TEST=Ran on lenoa (no efs2)

Change-Id: I029597e5ae4a9a432486ec37a282cabb6f5d53d0