lacros-au: Fix screen record bug
Lacros AU test is expected to save a screen record only
when it has an error in verifying versions.
If test passes, the record should be stopped and removed from
storage.

There was a bug that record is not deleted on test pass.

BUG=chromium:1258664
TEST=Run `tast run lacros.Update*`
TEST=Then `watch ls /home/chronos/user/Downloads` on DUT
TEST=to see if record is deleted whether test passed or failed

Change-Id: I7e0a719dac553a2ead179105fa25d82e45803c50
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3314697
Tested-by: Hyung Tae Kim <hyungtaekim@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Hyung Tae Kim <hyungtaekim@chromium.org>