faft: Support DUT-less testing
This is a bit of a hack.

Add a noSSH var to the firmware fixture, if set to true, then forbid SSH
calls to the DUT, and skip rebooting to the requested mode, and don't
change any GBB flags.

Change Helper functions that are very common in tests but use SSH, when
the DUT is not available.

EnsureDUTBooted changed to not ssh to DUT.

WaitConnect, will wait for S0 and sleep 20s.

DisconnectDUT becomes a NOOP.

Add OverridePlatform() since we can't ssh into the DUT to detect the
board and model.

Handle missing keyFile & keyDir in RequireServo.

BUG=b:208690862
TEST=tast run --var servo=${SERVO?}:nossh \
--var firmware.hibernate_time=1m --var noSSH=true \
--var board=grunt --var model=treeya 192.168.1.78 \
'("group:firmware" && firmware_bringup)'

Change-Id: I43c18286c44f787dfa7f95d43b1af14d390e191d