tast-tests: Fix the order of TestAPIConn call
`defer faillog.DumpUITreeOnError` was called before the
`defer closeBrowser(cleanupCtx)` which caused the UI Tree to be
recorded after the browser was already closed. Because of this
it's not possible to debug some test failures.

Change the order to get the better UI tree.

BUG=chromium:1268351
TEST=tast run <ip> accountmanager.*

Change-Id: I4e127b092e06b60a5c6d32accef5d33d35cddb36
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3310912
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Tested-by: Anastasiia N <anastasiian@chromium.org>
Commit-Queue: Anastasiia N <anastasiian@chromium.org>