Report power measurement duration as metric
The lacros performance tast test use a fixed interval of 20 seconds to
make comparative measurements. However, the implementation of this
interval appears to have some difficulty in terms of accuracy. Local
runs indicate at least a 8% variance in test interval time.

Here I am adding a metric to see how inaccurate this interval duration
enforcement is on the testing farm.

This is important because all of our testing is done as total energy in
joules so a longer test time will report as consuming more power.

BUG=chromium:1263337
TEST=tast.lacros.GpuCUJ.maximized_delegated

Change-Id: Ibc76766cf59103d6a0972f4085030a3ce4fa0a8c