video: enable ChromeStackDecoding everywhere
video.ChromeStackDecoding only runs on devices with the direct
VideoDecoder enabled. This CL enables it for all devices, since
eventually we want to migrate them all, and this way we get to
see the status everywhere.

I verified on kevin and hana that running ChromeStackDecoding
works as expected and, more importantly, doesn't break the
subsequent tests (in this case for example video.Contents was
still fine).

BUG=b:205354844
TEST=tast run -checktestdeps=false <IP> video.C* on kevin, hana

Change-Id: I072a3e8a5bbd3d9ccfeb461518e24a530c0011e5