tast-tests: Ignore spurious text shares in appsplatform.WebAPK
appsplatform.WebAPK is currently about 10% flaky due to an issue where
a share is received with no files attached.

This text share seems to happen because something on the Android side is
triggering an extra text share during the first part of the test. It's
not clear where this share comes from, but I've been unable to reproduce
the issue on a real device.

This CL works around the issue by just receiving and ignoring the text
share. The expected file share is then able to come through correctly
afterwards.

BUG=chromium:1254586
TEST=../platform/tast-tests/tools/repeat-test.sh -s -r 15 --
    $DUT appsplatform.WebAPK

Change-Id: I249ab7035e6c3ba5a0fa57be3bbff15eec32ee59