video cuj: dismiss notifications prompt before switching window
It has been seen that YouTube occasionally pops up alert dialog asking
if users want to enable notifications. If the alert dialog is not
dismissed before switching window to the targeted display, the operation
will fail as the current active window cannot be found. This commit
resolves the issue by moving the clearNotificationPromps procedure to
before the SwitchWindowToDisplay is called.

BUG=b:211055357
TEST=tast run <vars> <dut> ui.VideoCUJ2.basic_youtube_web,
plus_youtube_web

Change-Id: I813173b1d1f1344db0ca364557871617b0056202