tast-tests: Reduce CPU temp and opt-in pre-condition fails for arc.*Perf
Adjusted CPU temp threshold based on latest arc.AppLoadingPerf and
arc.AuthPerf statinless data to mitigate most of related failures.

For arc.AppLoadingPerf, opt-in is not required and thus will remove
it to fix pre-condition failures due to opt-in. This can not be
removed for arc.AuthPerf as it's a requirement for this tast.

Replace cpu.WaitUntilCoolDown for cpu.WaitUntilStabilized added from
https://crrev.com/c/3229351 to also check for CPU idle in arc.AuthPerf.

BUG=b:197685689
TEST=tast run $DUT arc.AppLoadingPerf and arc.AuthPerf

Change-Id: I4fc01edf6455f99d267931514a502d352f270756
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3321302
Reviewed-by: David Welling <davidwelling@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Tested-by: Alan Ding <alanding@chromium.org>
Commit-Queue: Alan Ding <alanding@chromium.org>