tast: Support new enterprise management protos package
All current protobuf files exist in Chromium are now supporting
go_package in their .proto definition.

go_package means that's the exact generated package name that
will be needed to be imported whenever there is a need to use
the the *.go protobuf generated files.

There is an ongoing migration to use the newer version of
protobuf. In order to do so the following should happen:
1) All chromium protobuf files have go_package defined in *.proto
2) ChromeOS picking up these new changes for all *.proto files
3) All dependent proto files (importing others from different
package) should be compiled correctly (i.e. adjusted from common-mk)

This CL is adjusting the used enterprise main proto package to be
enterprise_management_proto whenever it will be used in any
*.go file.

BUG=chromium:1164401
TEST=emerge-eve chromeos-base/tast-remote-tests-cros

Cq-Depend: chromium:3320513, chromium:3314595
Change-Id: I7091c6e01f60a7fda05e96eb7052d6e122d24065