tast-tests: Speculative fix for wmp.OverviewScroll.
Attempts to fix two issues.

Error with window not being visible. Probably the addition of the
playstore, added a wait for visible to address this.

On some boards, the windows are not offscreen. Probably because
scrolling can vary depending on the touchscreen of the DUT. Changes
swipes to be flings which scrolls more.

BUG=chromium:1261266
TEST=tast -verbose run <IP> wmp.OverviewScroll.*

Change-Id: Ibca05f71099f8b972c208b41a900fccbc262b142