faft: Change start mode recovery mode reboot tests
The tests firmware.BootMode.rec_{warm,cold} boot to recovery mode, do a
cold or warm reboot, and then verify the DUT is in normal mode. However
DUT was in dev mode before the test started, it will end in dev mode
instead of normal. To prevent this, use the fixture to ensure the DUT is
in normal mode, then reboot to recovers, then cold or warm reboot.
(whew)

BUG=b:169333650
TEST=firmware.BootMode.rec_cold firmware.BootMode.rec_warm on treeya
and kindred, starting in dev mode.

Change-Id: Ifb8e8843b63db44fb2d0cfdf06db9d9d177e1654
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3324296
Reviewed-by: Kshitij Shah <tij@google.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>