Modify holding space download tests to verify initial state.
When a download is first initiated, holding space UI should show
a prompt and chip in the pinned files section to educate the
user about pinning a file.

Note that this CL also increases the size of download files to reduce
test flakiness until a better means of throttling download speed can
be identified.

BUG=chromium:1251786
TEST=tast -verbose run <DUT> holdingspace.Download.*

Change-Id: I5a01540536303c1b0dc9284662034507c1100ae0