Add more allowed error entries for network.ShillNoErrorsInLog Tast test.
Our goal is to make this test pass at first. This involves
identifying all the error logs with the current state. Then we will
clean them up over time until there is no error logs.

After the initial addition of the test, I found three more error
entries that cause test failure. This is expected because the initial
allowed entries are based on the test on my DUT only.

BUG=b:204365880
TEST="$ tast run localhost:2222 network.ShillNoErrorsInLog" after SSH remote port forwarding

Change-Id: I7ed4bb6ac3709e455d3007e3223c581a4f5e2af3