Rewrite ui.ChromeCrashLoop from scratch
Rewrite ui.ChromeCrashLoop from scratch. Named rewrite "V2" until it can
be stabilized and replace the existing one. Intention is to completely
remove the current ui.ChromeCrashLoop at some point.

The rewrite just uses a command line flag to force Chrome to crash each
time it starts up, instead of sending a signal to it externally. This
makes the test:

1. simpler
2. less dependent on chromecrash (to start untangling
   go/the-mess-which-be-killandgetcrashfiles)
3. More resilient to unstable Chrome instances because now if Chrome crashes,
   it doesn't matter why. Hopefully, this will avoid us sherriffs disabling us
   because Chrome was unstable, like https://crbug.com/1276549

BUG=b:202795944
TEST=tast -verbose run --extrauseflags chrome_internal localhost:2222 ui.ChromeCrashLoopV2.* ui.ChromeValidity

Change-Id: I7220f9617d06e69fd5ff1747b39824ca57bccacd