u2fd: Add WebauthnU2fMode integration test
For WebAuthn, we currently have three ways of authenticating in Chrome
OS platform authenticator: PIN, fingerprint, and power button press. But
for now we only have an integration test for PIN authentication. In u2f
mode (enabled in managed devices), whether the user will be prompt for
PIN or power button authentication is determined by the combination of
the "user verification" and "authenticator type" field in the WebAuthn
request. Add the test WebauthnU2fMode to test that our Chrome OS
platform authenticator prompts for the correct authentication method
(PIN or power button), and the authentication succeeds.

Fix a typo in the IntegratedU2f test.

BUG=b:184374570
TEST=tast run -var=servo=$SERVO $DUT u2fd.WebauthnU2fMode

Change-Id: Ib4397e9b05a2d3c413afa2775e505d33b210f5d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3329442
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>