Fix Camera Roll Tast test
Camera Roll was failing to load because it was unable to generate
thumbnail for the android-screen.mp4 file that is created for each
test run. The content of that file was removed at the end of each
test but it wasn't removed from the phone's gallery, which confused
Camera Roll.

Updated the test code to use the new RemoveMediaFile function to
properly clean up that file.

BUG=b:194228922
TEST=this

Change-Id: I69abee430b2b92323c2a7d3acfa4886772aa93f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3329797
Tested-by: Jason Sun <jasonsun@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>