arc: Temporarily ignore verification for SELinux for "misc" dir
CL on rvc-arc branch ag/q/topic:perfetto-configs-rvc-arc blocked
by this tast test due to add a new file context. Although it change
the prebuilts of selinux as well, before PFQ the change cannot be
landed and be read from tast test. So after the CLs landed, we
will revert this change.

BUG=b:208933312
TEST=tast run <DUT> arc.SELinuxFilesDataDir.vm

Change-Id: I7d5b1402029c53478b8b5ab27827cce2da5ce1b0