arc: Temporary ignore "misc" directory for land selinux change
CL on rvc-arc branch ag/q/topic:perfetto-configs-rvc-arc
blocked by the current tast-test as it adds a new file context.

After the CL is landed, the plan is to add the appropriate
SELinux context in the source for correct verification by this tast-test.

Post that, the current CL will be reverted.

BUG=b:208933312
TEST=tast run <DUT> arc.SELinuxFilesDataDir.vm

Change-Id: I7d5b1402029c53478b8b5ab27827cce2da5ce1b0