tast-tests: Update filemanager.ZipPerf to run for SWA
Files app is migrating from a Chrome app to System Web App and in the
process we want to evaluate the performance impact of the migration on
the Zip subsystem. To do this introduce a new test parameter that
evaluates the performance entirely on the SWA version of Files app.

BUG=b:207576612
TEST=tast run filemanager.ZipPerf*

Change-Id: Ic8f34a4ee468beb8d38c2b747ba6844c71303cdc