Stabilize ui.LauncherAnimationPerf.lacros.
Looks like timing issue.
The original code didn't wait for Lacros to show an initial about:blank
window, so trying to close it may happen before the window is shown.

This CL fixes it by changing the set up of the test.
With this CL, for each iteration, all windows are closed, then a new
tabs are opened, with properly waiting for lacros's initial about:blank tab.

BUG=chromium:1276072
TEST=Ran ui.LauncherAnimationPerf.lacros.

Change-Id: Ic8d2572da2bb39c763bd242679a0fa9e79f53d09