lacros: Don't write GPUCUJ traces to disk.
Something (currently unknown) is causing /tmp to become full. This in
turn causes Lacros tests to fail, which somehow cascades into non-Lacros
tests failing. By not writing Lacros CUJ traces to disk, we think this
will put us again slightly under the boundary of /tmp being full, which
will allow tests to pass again.

This fix is a bandaid and is not intended to be a long term solution.
See bug for details. Another option that was considered was disabling
the Lacros GPUCUJ tests altogether. That would change the way that tests
get sharded and could itself induce more failures. The current approach
was deemed the safest option.

BUG=chromium:1277606
TEST=fastbuild.sh -T

Change-Id: I94e207930b97ab0f0030894b2f57f61ad6769537