tast-tests: Fix ShareMovies tests flaky failures
The ShareMovies test starts the PlayStore app and waits for it to be
ready first, and then opens the Files app to share the Movies
subdirectory with Crostini. When the PlayStore app finishes
setting things up, it restarts itself in the end, which takes a bit of
time for the old PlayStore window to be closed and the new one being
opened.

At present, the ShareMovies test does not wait for the PlayStore app to
be restarted and just continues with the rest of the test, by opening
the Files app and attempting to open the "Play files" directory. As a
result, the new PlayStore window (after it restarts) may sometimes open
on top of the Files app and cause the remaining test actions to fail.

Make sure the test waits for the PlayStore app to restart itself, and
then close the PlayStore app as it is no longer required anyway.

BUG=b:198565709, b:198565474, b:198565763, b:199243643
TEST="tast run $DUT crostini.ShareMovies.*"

Change-Id: I90cb39feee6152f4ec291556ec77c0b7395d8b81