Use /usr/local/tmp/lacros* directory as fallback user data directory for lacros.
Currently, the user data directory is created under the runtime directory.
However, if rootfs lacros is used, because it is read only,
we cannot create the directory under it, so as a fallback we have created
the user data directory in /tmp.

However, this causes an issue, because /tmp is memory mapped and its size
is limited.
This CL moves the fallback location to /usr/local/tmp so it should be
resolved.

BUG=chromium:1277606
TEST=tryjob.

Change-Id: Iaa1da2b74571e0e1f0dc473859dc1c05ff6f9e66