tast tests: Fix the test for new GMSCore UI.
The test has started failing as there is new UI
change in GMSCore. But sometimes the Old UI is displayed
and hence tests fails. There is no straight forward way
to turn off the flag to turn this UI change. Hence fixing
test with flag to make it work for both old and new UI

BUG=b:204823538
TEST=repeat-test.sh -s -r 30 <DUT> arc.VerifySettings.vm

Change-Id: Ia8069f11544949775c8a088567e944dd3c528f1d