arcappcompat: Update tests to use arcBooted fixture
Update the standardized tests to use the arcBooted fixture, since it is
stable enough to be used in 'critical' tests, and set the device mode
within each test run. Note that the
mouse_scroll.resize_lock_smooth_scroll_vm test still requires the
arcBootedWithTouchModeMouse fixture as it requires specific features to
be enabled.

BUG=b:194900373
TEST=push apk && tast run <DUT> arc.Standardized*

Change-Id: I3bf1c483383ef168d1144a271782c82eb9e8cbed