arcappcompat: Fix touchview split screen failures on Krane
arcappcompat tests are failing in the touchview split screen
test case on Krane. It is observed that when the app is in
landscape orientation, the app window is snapping on the
bottom and top instead of left and right. Added a code to
ensure device display orientation is in landscape so that
the app window snaps on left and right.
Also observed that some of the apps are failing flaky
if their orientation type is “portrait-secondary“.
Added “portrait-secondary“ check, so that test would skip
on apps having portrait-secondary orientation.

BUG=b:202890874

TEST=tast run <krane> arcappcompat.GooglePhotos.tablet_mode
tast run <krane> arcappcompat.Asana.tablet_mode
tast run <helios> arcappcompat.GooglePhotos.vm_tablet_mode
tast run <eve> arcappcompat.GooglePhotos.vm_tablet_mode

Change-Id: I962ccd66185affda62e8501f2acd180e3f178de8