Modify variations smoke test to also run on non-Chrome branded builds.
I'm going to add this test to a VM builder in the CI, so it also needs
to run on Chromium builds. Unofficial builds do not ping the Finch
server by default, so explicitly specify the production server to cover
the case.

BUG=b:202325793
TEST=tast run <dut> ui.ChromeVariationsSmoke on an official build, and
on an amd64-generic VM without chrome branding

Change-Id: Ice4fb53c2adcd6a1d57e8f73c9af82c871cfad1a