tast-tests: Stabilize system account addition
The iframe that shows the Gaia sign-in screen is hidden by default and
is being shown after the webpage inside it is loaded. It looks like
sometimes the ui tree is not updated at the right moment and it thinks
that the iframe node is empty. Fix this by going to the previous screen
and navigating back to the iframe node. This will hide and show the node
and will make the ui tree update.

Add accountmanager.SystemDialog test to track the stability of the
account addition dialog separately from other tests.

BUG=chromium:1268351
TEST=tast run <ip> accountmanager.*

Change-Id: I826b53ebf932d6d68727b9b59ba7f2a6566780a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3333918
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Tested-by: Anastasiia N <anastasiian@chromium.org>
Commit-Queue: Anastasiia N <anastasiian@chromium.org>