wifi: SetTXPower: remove dependency
cros_config is still deployed on boards without the USE=unibuild flag
on (see
https://source.corp.google.com/chromeos_public/src/third_party/chromiumos-overlay/chromeos-base/chromeos-config-tools/chromeos-config-tools-9999.ebuild;rcl=d9fcbbbf141cec87dd6bf910cf81350e361cbffa;l=61),
and this test should be running on all boards anyway. Speculatively
remove the dependency (to avoid dependencies on OOO folks), monitor
tests, and file a bug to fix it if nothing breaks.

BUG=b:211487079
TEST=`tast run ${KEVIN_DUT} wifi.SetTXPower`

Change-Id: I92d0308fd7f35f8642b2625856a78667dedfa77b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3350715
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>