faft: Use wget instead of piping image to ssh
Instead of piping the data from CloudStorage to ssh to the tar command
on the servo host, get the url to the data, and use wget on the servo
host to get the data directly. This is much faster on my local machine,
and piping to a command doesn't work yet on docker servod.

BUG=b:207133139
TEST=tast run firmware.Fixture.rec locally

Change-Id: I20de9b1dd70f421ccea0af31347e46dfcab2abda