platform.BootPerf: Collect firmware timestamps from cbmem output
This change makes the test to get firmware boot time from the output of
`cbmem -t` instead of waiting and reading from /tmp/firmware-boot-time.
The full output of `cbmem -t` is also saved for debugging.

This is a port of the tauto counterpart changes:
https://crrev.com/c/2844963 and https://crrev.com/c/2762487

BUG=b:172221728
TEST=Ran on hatch

Change-Id: Iec5828eee4e5e2efd43461e209f4f7d697c91192
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3353075
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Reviewed-by: Pranav Batra <batrapranav@chromium.org>
Tested-by: Chinglin Yu <chinglinyu@chromium.org>
Commit-Queue: Chinglin Yu <chinglinyu@chromium.org>