quicksettings: Make NotificationsHidden return !exists
Until now NotificationsHidden() function checked if Notification Center
node exists and returned that information to caller. True if it exists.

Test quicksettings.LockScreen expects from the function to get
information if notifications are hidden. True if they are hidden (as
the name suggests).

If Notification Center node don't exist, the function will return false
which is interpreted by the test that notifications are not hidden,
which is wrong. The function should return true if Notification Center
don't exist.

BUG=chromium:1280807
TEST=tast -verbose run <DUT_IP> quicksettings.LockScreen

Change-Id: I2c24a73a9f32adc8562eb459d4be2e7372cdc932