[tast-test] Make launcher.AppDragAndDrop more stable
The test was dragging items on the first page around.
It needs at least 4 items on the page to work, which may not be
satisfied for certain deviced - the first page will generally contain
only some default apps with a page break before other apps. While
testing on kukui, that ammounted to 3 apps on the first page.
To address this, change the test to preinstall at 100 apps, and make it
drag items from the second page (which should be full, and have enough
items).

Also, after creating a folder, the test was attempting to open the
folder using LeftClick, which was failing because the folder item
location seemed to be stale - use FocusAndWait before attempting click,
which should force automation layer to update the folder location.

BUG=chromium:1282319
TEST=Ran launcher.* tests on kukui

Change-Id: I00f1e1c573805be071d944dd488eab8452e732ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3355411
Reviewed-by: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Matt Stark <msta@google.com>
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Tested-by: Toni Barzic <tbarzic@chromium.org>