[tast-test] Make launcher.AppDragAndDrop more stable
The test was dragging items on the first page around.
It needs at least 4 items on the page to work, which may not be
satisfied for certain devices - the first page will generally contain
only some default apps with a page break before other apps. While
testing on kukui, that amounted to 3 apps on the first page.
To address this, change the test to preinstall 100 apps, and make it
drag items from the second page (which should be full, and have enough
items).

Also, after creating a folder, the test was attempting to open the
folder using LeftClick, which was failing because the folder item
location seemed to be stale - use FocusAndWait before attempting click,
which should force automation layer to update the folder location.

BUG=chromium:1282319
TEST=Ran launcher.* tests on kukui

Change-Id: I00f1e1c573805be071d944dd488eab8452e732ae