Refactor Activity.Start() to use functional option pattern
This will allow us to remove the StartWithArgs method and just
optionally pass all the functional options that are appropriate. It will
also help tast developers by making it more obvious what options are
available to them for the start command.

Also add a StartWithDefaultOptions method which starts the activity with
the old default args that the bare Start method would run with.

Change all tast tests to use the new Start or StartWithDefaultOptions
method syntax.

BUG=b:203214749
TEST=tast run $DUT arc.SurfaceOrientation and observe full screen windowing
mode appear.
TEST=CQ Dry run will run all the tests that are affected

Cq-Depend: chrome-internal:4474024
Change-Id: Iba0503dd75c591d5f8303a21f8f45e7f0aafeeb8