Activity.StartWithOptions to standardize am start args
Add a member function of arc.Activity that allows you to start an
activity with an instance of ActivityStartOptions instead of a list of
prefix and suffix args in string form.

Since the new ActivityStartOptions struct has a standardized set of
fields which developers can set, this helps developers be more
consistent when writing tests that use activities.

Add unit tests for the ActivityStartOptions struct and its methods.

Have the arc.SurfaceOrientation test use the the new start function.

BUG=b:203214749
TEST=tast run $DUT arc.SurfaceOrientation and observe full screen windowing
mode appear.
TEST=Run the new unit test

Change-Id: I8a2ea96ce72b6e6b579b15ff3df5f2d7c8cb22dd