tast-tests: Fix arc.FileSystemPermissions gid check failure
From auto-bisector, culprit https://crrev.com/f4d137c61638f1277ac097d964075bb364ec23aa was found to break the tast in 14365.0.0.  Seem like
the CL is benign and we just have to update the expectation of the test
to match the change.

Manually verified chrome://arc-overview-tracing and chrome://arc-graphics-tracing on caroline/R98-14375.0.0.

BUG=b:209034160
TEST=tast run $DUT arc.FileSystemPermissions

Change-Id: I76434cf45330b450abea560770f5d29880805dc2