Add external facing grpc svc ChromeService
It is part of the Exposing Tast Services through gRPC initiative.
The service is intended for use outside of Tast (e.g. Mobly).
For more information, see http://go/tast-tcp-grpc-svc

This CL includes:
1) CrosServerUtil that handles the cros server lifecycle and abstracts
that from the end users.
2) SharedObjectsForServices singleton provides shared access to common
objects such as chrome across services.  See http://shortn/_ZENds5xwe9
3) ChromeService provides functionalities for creating and closing
chrome sessions
3.5) Reclaim chrome_service.proto for ChromeService. New restrictions
are imposed to prevent protos with the same name (even from different
packages) from being registered.  chrome_services.proto has been used
as an internal example on how to interact with the SSHPipe based
gRPC server. Since the new service is going to be public, I feel that
it would be better for the names to be as intuitive/simple as possible.
4) Remote test for ChromeService

BUG=b:208267010
TEST=tast run -build=true <DUT> *GRPC*
including browser.ChromeServiceGRPC.* and example.GRPC

Change-Id: I89c196b24b642234e048cdb18a9ac2fae1994113