tast-tests: Support WiFi6e and add WiFi6e tests
Add support to the GT-AXE11000 router. This is a WiFi6e router. Tests
are added to test WPA3 on the 6ghz band. OWE (6ghz open networks) are currently not supported
on ChromeOS devices so those tests are not included but the security
file has been included. We also test hidden networks but this is
currently non-functional.

BUG=b:204498033
TEST=tast run wifi.AxSimpleConnect.*

Change-Id: I9364af77ccb0293ff14f96867d24afc3f7fd9d11
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3358471
Reviewed-by: Kevin Lund <kglund@google.com>
Reviewed-by: Jun Yu <junyuu@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Billy Zhao <billyzhao@chromium.org>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>