tast-tests: Close all windows in test cleanup
Don't close window in IsAccountPresentInArc and close all open windows
in TestCleanup instead. This will allow to get ui_tree if the result of
IsAccountPresentInArc is not correct.

Currently ARC window is closed before we check the return value and
there is no way to see the screenshot of the window when the test fails.

BUG=chromium:1268351
TEST=tast run <ip> accountmanager.*

Change-Id: Ifd388b6d0f72306e6157c09ded57b61f78084165
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3365040
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Tested-by: Anastasiia N <anastasiian@chromium.org>
Commit-Queue: Anastasiia N <anastasiian@chromium.org>