crash: Wait for crasher, not crash_reporter
If we wait for crash_reporter to exit, we could spuriously fail if there
is another crash at approximately the same time as the crash that we
induce, and especially if the process running crash_reporter takes a
while to reap it (as happens with the JS error reporting pipeline from
chrome). So, instead of waiting for crash_reporter to exit, wait for the
*crasher* process, since it cannot be reaped until after crash_reporter
is finished processing it anyway.

BUG=None
TEST=tast run forward_crbook crash.User.*

Change-Id: I5887eb82c3f22cbe54e1accd754abcee0b8e745a