arcappcompat: Update Tast Test for Messenger
BUG=b:211154669
TEST=tast run <eve> arcappcompat.Messenger.vm_clamshell_mode
     Tast run <kevin> arcappcompat.Messenger.tablet_mode

Messenger is failing in launch due to a change in UiObject Id of
username. Added a regex to fix it. On some devices, observed a
new dialog box is prompted to save the password. Added a code
to handle it. Also replaced the homepage verifier with a launch
verifier to avoid flaky launch failures.

Change-Id: I42ad9bc3cd15d559131ceb574093874bf050dcc1