uidetection: revise the use of options
Moved the options to the attribute of the uidetection context.

This allows the test to choose its default options of its own.
Previously, if a test needs to retry 10 times for each action,
it needs to add it for all of them like:
uda.LeftClick(...., Retry(10)),
uda.LeftClick(...., Retry(10)),
uda.LeftClick(...., Retry(10))

After this change, the workflow is simplified as:
uda:=uda.WithOptions(Retry(10))
uda.LeftClick(....)
uda.LeftClick(....)
uda.LeftClick(....)

BUG=None
TEST=tast run <DUT> uidetection.BasicDetections

Change-Id: I93a5339379d191caa728beb972960754d7591e1f