camera: Gives more informative error message when timeout
Originally, capture unittests always output:
"Failed to run capture_unittests: context deadline exceeded"
when hitting timeout, which is not very informative.

This CL tries to parse the log and get the last test case when hitting
timeout so at least we might be able to understand which test case does
it stuck at from the error message.

BUG=b:210084868
TEST=tast run [DUT] camera.CaptureUnittests

Change-Id: I7b7aafd1776b95636215814dc33aa544a4343bcd