storage: Parse the life time information for eMMC devices
eMMC devices report type A and type B lifetime estimates, the values
represent a range of percentage used values that the device falls in.
We need to report at least the larger of the 2 values (they are determined
by the vendor) to get an idea of the lifetime progression during v2 quals.

BUG=b:213650928
TEST=Ran storage.FullQualificationStress.stress to confirm data is now correctly
reported, ran ./fast_build.sh -v -t chromiumos/tast/common/storage

Change-Id: I07c7e8e6bc2d5b74b267b256f50261b843880bee